Fix H-pulling?
In retrospect the answer is easy, but man this took a while.
