+++ Content.Server/Pulling/PullingSystem.cs
@@ -1,6 +1,10 @@
-﻿using Content.Shared.Pulling;
+using Content.Shared.Input;
+using Content.Shared.Pulling;
+using Content.Shared.Pulling.Components;
 using JetBrains.Annotations;
 using Robust.Server.GameObjects;
+using Robust.Shared.Input.Binding;
+using Robust.Shared.Players;
 
 namespace Content.Server.Pulling
 {
@@ -15,6 +19,32 @@ namespace Content.Server.Pulling
 
             SubscribeLocalEvent<PullableComponent, PullableMoveMessage>(OnPullableMove);
             SubscribeLocalEvent<PullableComponent, PullableStopMovingMessage>(OnPullableStopMove);
+
+            CommandBinds.Builder
+                .Bind(ContentKeyFunctions.ReleasePulledObject, InputCmdHandler.FromDelegate(HandleReleasePulledObject))
+                .Register<PullingSystem>();
+        }
+
+        private void HandleReleasePulledObject(ICommonSession? session)
+        {
+            var player = session?.AttachedEntity;
+
+            if (player == null)
+            {
+                return;
+            }
+
+            if (!TryGetPulled(player, out var pulled))
+            {
+                return;
+            }
+
+            if (!pulled.TryGetComponent(out SharedPullableComponent? pullable))
+            {
+                return;
+            }
+
+            pullable.TryStopPull();
         }
     }
 }
