+++ Content.Shared/Pulling/Systems/SharedPullingSystem.cs
@@ -62,7 +62,6 @@ namespace Content.Shared.Pulling
 
             CommandBinds.Builder
                 .Bind(ContentKeyFunctions.MovePulledObject, new PointerInputCmdHandler(HandleMovePulledObject))
-                .Bind(ContentKeyFunctions.ReleasePulledObject, InputCmdHandler.FromDelegate(HandleReleasePulledObject))
                 .Register<SharedPullingSystem>();
         }
 
@@ -192,28 +191,6 @@ namespace Content.Shared.Pulling
             pullable.TryMoveTo(coords.ToMap(EntityManager));
 
             return false;
-        }
-
-        private void HandleReleasePulledObject(ICommonSession? session)
-        {
-            var player = session?.AttachedEntity;
-
-            if (player == null)
-            {
-                return;
-            }
-
-            if (!TryGetPulled(player, out var pulled))
-            {
-                return;
-            }
-
-            if (!pulled.TryGetComponent(out SharedPullableComponent? pullable))
-            {
-                return;
-            }
-
-            pullable.TryStopPull();
         }
 
         private void SetPuller(IEntity puller, IEntity pulled)
